backend/requirements.txt
# TURBO-PLAN X - Backend Requirements

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Logging
loguru==0.7.2

# WebSocket
websockets==12.0

# CORS
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
